+++ docs/changelog.rst
@@ -13,6 +13,7 @@ Added
 -----
 - Before writing the forcing to disk, throw an error for any missing values (#748)
 - Added upgrade support for the latest Wflow.jl v1.1 (#765)
+- Added ability for users to exclude reservoirs that are outside of the river network, using `exclude_outside_reservoirs` flag. (#777)
 
 Fixed
 -----